fix: replaced float with int return

+++ includes/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: mpouillo <mpouillo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/13 12:15:41 by mpouillo          #+#    #+#             */
-/*   Updated: 2025/12/13 16:02:26 by mpouillo         ###   ########.fr       */
+/*   Updated: 2025/12/14 11:42:56 by mpouillo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,6 +37,6 @@ typedef struct s_stack
 t_stack	*stack_create(char **item_list, int size);
 void	stack_push(t_stack *stack, int item);
 int		stack_pop(t_stack *stack);
-float	compute_disorder(t_stack a);
+int		compute_disorder(t_stack a);
 
 #endif
